fix sign compare warnings on NO_STREAM_ID_FILTER
The NO_STREAM_ID_FILTER is defined as ~0U and that is unsinged.

Make all checks unsigned to make compiler happy on the following cases.

../../git/lib/dvb/frontend.cpp:2630:34: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   bool multistream = (parm.is_id != NO_STREAM_ID_FILTER || (parm.pls_code & 0x3FFFF) != 1 ||
                                  ^
../../git/lib/dvb/frontendparms.cpp:345:16: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  if (stream_id == NO_STREAM_ID_FILTER) return transponderParameters.is_id;
                ^

+++ lib/dvb/frontend.cpp
@@ -2627,7 +2627,7 @@ int eDVBFrontend::isCompatibleWith(ePtr<iDVBFrontendParameters> &feparm)
 		{
 			return 0;
 		}
-		bool multistream = (parm.is_id != NO_STREAM_ID_FILTER || (parm.pls_code & 0x3FFFF) != 1 ||
+		bool multistream = (static_cast<unsigned int>(parm.is_id) != NO_STREAM_ID_FILTER || (parm.pls_code & 0x3FFFF) != 1 ||
 					(parm.pls_mode & 3) != eDVBFrontendParametersSatellite::PLS_Root);
 		if (parm.system == eDVBFrontendParametersSatellite::System_DVB_S2 && multistream && !is_multistream())
 		{
